re-created model and re-ran migration to make sure all the files and directories are in the right place
