+++ models/index.js
@@ -1,23 +1,12 @@
-"use strict";
+'use strict';
 
-const http = require("http");
-const hostname = "127.0.0.1";
-const port = 3000;
-const express = require("express");
-const app = express();
-const server = http.createServer(app);
-const fs = require("fs");
-const path = require("path");
-const { Sequelize, Model, DataTypes } = require("sequelize");
+const fs = require('fs');
+const path = require('path');
+const Sequelize = require('sequelize');
 const basename = path.basename(__filename);
-const env = process.env.NODE_ENV || "development";
-const config = require(__dirname + "/../config/config.json")[env];
+const env = process.env.NODE_ENV || 'development';
+const config = require(__dirname + '/../config/config.json')[env];
 const db = {};
-const bodyParser = require("body-parser");
-
-app.use(express.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
 
 let sequelize;
 if (config.use_env_variable) {
@@ -26,16 +15,17 @@ if (config.use_env_variable) {
   sequelize = new Sequelize(config.database, config.username, config.password, config);
 }
 
-fs.readdirSync(__dirname)
-  .filter((file) => {
-    return file.indexOf(".") !== 0 && file !== basename && file.slice(-3) === ".js";
+fs
+  .readdirSync(__dirname)
+  .filter(file => {
+    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
   })
-  .forEach((file) => {
+  .forEach(file => {
     const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
     db[model.name] = model;
   });
 
-Object.keys(db).forEach((modelName) => {
+Object.keys(db).forEach(modelName => {
   if (db[modelName].associate) {
     db[modelName].associate(db);
   }
@@ -45,94 +35,3 @@ db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
 module.exports = db;
-
-class User extends Model {}
-
-User.init(
-  {
-    user_name: DataTypes.STRING,
-    first_name: DataTypes.STRING,
-    last_name: DataTypes.STRING,
-    email: DataTypes.STRING,
-    password: DataTypes.STRING,
-    skill_level: DataTypes.STRING,
-  },
-  {
-    sequelize,
-    modelName: "User",
-  }
-);
-
-// get all users
-app.get("/users", async (req, res) => {
-  res.setHeader("Content-Type", "application/json");
-  const users = await User.findAll();
-  console.log(users);
-  res.status(200).send(users);
-  //console.log(users);
-});
-
-// get one user
-app.get("/users/:id", async (req, res) => {
-  res.setHeader("Content-Type", "application/json");
-  let userId = req.params["id"];
-  const users = await User.findAll({
-    where: { id: userId },
-  });
-  res.status(200).send(users);
-  //console.log(users);
-});
-
-// add a user
-app.post("/users", async (req, res) => {
-  res.setHeader("Content-Type", "application/json");
-  await User.create({
-    user_name: req.body.user_name,
-    first_name: req.body.first_name,
-    last_name: req.body.last_name,
-    email: req.body.email,
-    password: req.body.password,
-    skill_level: req.body.skill_level,
-  });
-
-  res.status(200).send("User added");
-  //console.log(users);
-});
-
-// update a user
-app.put("/users/:id", async (req, res) => {
-  res.setHeader("Content-Type", "application/json");
-  let userId = req.params["id"];
-  const users = await User.update(
-    {
-      user_name: req.body.user_name,
-      first_name: req.body.first_name,
-      last_name: req.body.last_name,
-      email: req.body.email,
-      password: req.body.password,
-      skill_level: req.body.skill_level,
-    },
-    {
-      where: {
-        id: userId,
-      },
-    }
-  );
-  res.status(200).send("User updated");
-  //console.log(users);
-});
-
-// delete a user
-app.delete("/users/:id", async (req, res) => {
-  res.setHeader("Content-Type", "application/json");
-  let userId = req.params["id"];
-  const users = await User.destroy({
-    where: { id: userId },
-  });
-  res.status(200).send("User was deleted");
-  //console.log(users);
-});
-
-server.listen(port, hostname, () => {
-  console.log(`Server running at http://${hostname}:${port}/`);
-});
